\name{toyrpkg-package}
\alias{toyrpkg-package}
\alias{toyrpkg}
\docType{package}
\title{
A toy R package 
}
\description{
A toy R package for study, demo, and test
}
\details{
\tabular{ll}{
Package: \tab toyrpkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-24\cr
License: \tab GPL=3\cr
}
A toy R package for study, demo, and test R's functionality, 
how to write packages for R, etc. 
}
\author{
Mav
}
\keyword{package}
