\name{toyRpkg-package}
\alias{toyRpkg-package}
\alias{toyRpkg}
\docType{package}
\title{
Toy R package: a series of examples to demonstrate how to write a R package  
}
\description{
Demonstrate how to write R packages.  
}
\details{
\tabular{ll}{
Package: \tab toyRpkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-14\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: MG <mg203@users.sf.net> 
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# no examples now 
}
